requirements.txt
cachetools==5.3.1
fastapi==0.101.1
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
